blocked portal when using nixos

+++ private_dot_config/hypr/executable_desktop-portal.sh
@@ -1,4 +1,7 @@
 #!/bin/sh
+if [ -f /etc/NIXOS ]; then
+	exit
+fi
 sleep 1
 killall xdg-desktop-portal-hyprland
 killall xdg-desktop-portal-wlr
